+++ httpclient-jetty/src/main/java/io/fabric8/kubernetes/client/jetty/JettyWebSocket.java
@@ -21,28 +21,35 @@ import io.fabric8.kubernetes.client.http.HttpRequest;
 import io.fabric8.kubernetes.client.http.WebSocket;
 import io.fabric8.kubernetes.client.http.WebSocketResponse;
 import io.fabric8.kubernetes.client.http.WebSocketUpgradeResponse;
+import io.fabric8.kubernetes.client.utils.Utils;
 import org.eclipse.jetty.websocket.api.Session;
 import org.eclipse.jetty.websocket.api.UpgradeResponse;
 import org.eclipse.jetty.websocket.api.WebSocketListener;
 import org.eclipse.jetty.websocket.api.WriteCallback;
+import org.eclipse.jetty.websocket.api.exceptions.CloseException;
 import org.eclipse.jetty.websocket.api.exceptions.UpgradeException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
+import java.net.ProtocolException;
 import java.nio.ByteBuffer;
 import java.nio.channels.ClosedChannelException;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.locks.Condition;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
 public class JettyWebSocket implements WebSocket, WebSocketListener {
+
+  private static final Logger LOG = LoggerFactory.getLogger(JettyWebSocket.class);
+
   private final WebSocket.Listener listener;
   private final AtomicLong sendQueue;
   private final Lock lock;
   private final Condition backPressure;
-  private final AtomicBoolean closed;
+  private final CompletableFuture<Void> terminated = new CompletableFuture<>();
   private boolean moreMessages;
   private volatile Session webSocketSession;
 
@@ -51,13 +58,12 @@ public class JettyWebSocket implements WebSocket, WebSocketListener {
     sendQueue = new AtomicLong();
     lock = new ReentrantLock();
     backPressure = lock.newCondition();
-    closed = new AtomicBoolean();
     moreMessages = true;
   }
 
   @Override
   public boolean send(ByteBuffer buffer) {
-    if (closed.get() || !webSocketSession.isOpen()) {
+    if (terminated.isDone() || !webSocketSession.isOpen()) {
       return false;
     }
     buffer = BufferUtil.copy(buffer);
@@ -67,6 +73,10 @@ public class JettyWebSocket implements WebSocket, WebSocketListener {
       @Override
       public void writeFailed(Throwable x) {
         sendQueue.addAndGet(-size);
+        if (webSocketSession.isOpen()) {
+          LOG.warn("Queued write did not succeed", x);
+        }
+        webSocketSession.disconnect(); // prevent further writes
       }
 
       @Override
@@ -78,12 +88,24 @@ public class JettyWebSocket implements WebSocket, WebSocketListener {
   }
 
   @Override
-  public boolean sendClose(int code, String reason) {
-    if (webSocketSession.isOpen() && !closed.getAndSet(true)) {
-      webSocketSession.close(code, reason);
-      return true;
+  public synchronized boolean sendClose(int code, String reason) {
+    if (!webSocketSession.isOpen()) {
+      return false;
     }
-    return false;
+    webSocketSession.close(code, reason, new WriteCallback() {
+      @Override
+      public void writeFailed(Throwable x) {
+        LOG.warn("Queued close did not succeed", x);
+        webSocketSession.disconnect(); // immediately terminate
+      }
+
+      @Override
+      public void writeSuccess() {
+        CompletableFuture<Void> future = Utils.schedule(Runnable::run, webSocketSession::disconnect, 1, TimeUnit.MINUTES);
+        terminated.whenComplete((v, ignored) -> future.cancel(true));
+      }
+    });
+    return true;
   }
 
   @Override
@@ -118,7 +140,7 @@ public class JettyWebSocket implements WebSocket, WebSocketListener {
 
   @Override
   public void onWebSocketClose(int statusCode, String reason) {
-    closed.set(true);
+    terminated.complete(null);
     listener.onClose(this, statusCode, reason);
   }
 
@@ -130,7 +152,8 @@ public class JettyWebSocket implements WebSocket, WebSocketListener {
 
   @Override
   public void onWebSocketError(Throwable cause) {
-    if (cause instanceof ClosedChannelException && closed.get()) {
+    boolean completed = terminated.complete(null);
+    if (cause instanceof ClosedChannelException && !completed) {
       // TODO: Check better
       //  It appears to be a race condition in Jetty:
       // - The server sends a close frame (but we haven't received it)
@@ -140,7 +163,10 @@ public class JettyWebSocket implements WebSocket, WebSocketListener {
       // - Jetty throws a ClosedChannelException
       return;
     }
-    listener.onError(this, cause, cause instanceof IOException);
+    if (cause instanceof CloseException) {
+      cause = new ProtocolException().initCause(cause);
+    }
+    listener.onError(this, cause);
   }
 
   private void backPressure() {