+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/AbstractWatchManager.java
@@ -22,6 +22,7 @@ import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.KubernetesResource;
 import io.fabric8.kubernetes.api.model.ListOptions;
 import io.fabric8.kubernetes.api.model.Status;
+import io.fabric8.kubernetes.api.model.StatusDetails;
 import io.fabric8.kubernetes.api.model.WatchEvent;
 import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.Watch;
@@ -36,14 +37,19 @@ import io.fabric8.kubernetes.client.utils.internal.SerialExecutor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
 import java.net.MalformedURLException;
+import java.net.ProtocolException;
 import java.net.URL;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 
 import static java.net.HttpURLConnection.HTTP_GONE;
@@ -88,12 +94,14 @@ public abstract class AbstractWatchManager<T extends HasMetadata> implements Wat
 
   public static class WatchRequestState {
 
-    private final AtomicBoolean reconnected = new AtomicBoolean();
-    private final AtomicBoolean closed = new AtomicBoolean();
+    final AtomicBoolean reconnected = new AtomicBoolean();
+    final AtomicBoolean closed = new AtomicBoolean();
+    final CompletableFuture<Void> ended = new CompletableFuture<>();
 
   }
 
   private static final Logger logger = LoggerFactory.getLogger(AbstractWatchManager.class);
+  private static final int INFO_LOG_CONNECTION_ERRORS = 10;
 
   final Watcher<T> watcher;
   final AtomicReference<String> resourceVersion;
@@ -110,7 +118,11 @@ public abstract class AbstractWatchManager<T extends HasMetadata> implements Wat
 
   private final boolean receiveBookmarks;
 
-  private volatile WatchRequestState latestRequestState;
+  volatile WatchRequestState latestRequestState;
+  private final Map<Class<?>, Integer> endErrors = new ConcurrentHashMap<>();
+  private AtomicInteger retryAfterSeconds = new AtomicInteger();
+
+  private int watchEndCheckMs = 120000;
 
   AbstractWatchManager(
       Watcher<T> watcher, BaseOperation<T, ?, ?> baseOperation, ListOptions listOptions, int reconnectLimit,
@@ -142,9 +154,27 @@ public abstract class AbstractWatchManager<T extends HasMetadata> implements Wat
    * If forceClosed has not been set, then it's expected that the watch will
    * attempt to reconnect
    */
-  public void closeRequest() {
-    Optional.ofNullable(latestRequestState).ifPresent(state -> state.closed.set(true));
-    closeCurrentRequest();
+  public synchronized void closeRequest() {
+    WatchRequestState state = latestRequestState;
+    if (state != null && state.closed.compareAndSet(false, true)) {
+      logger.debug("Closing the current watch");
+      closeCurrentRequest();
+      CompletableFuture<Void> future = Utils.schedule(Runnable::run, () -> failSafeReconnect(state), watchEndCheckMs,
+          TimeUnit.MILLISECONDS);
+      state.ended.whenComplete((v, t) -> future.cancel(true));
+    }
+  }
+
+  private synchronized void failSafeReconnect(WatchRequestState state) {
+    if (state == latestRequestState && !forceClosed.get() && (reconnectAttempt == null || reconnectAttempt.isDone())) {
+      logger.error("The last watch has yet to terminate as expected, will force start another watch. "
+          + "Please report this to the Fabric8 Kubernetes Client development team.");
+      reconnect();
+    }
+  }
+
+  public void setWatchEndCheckMs(int watchEndCheckMs) {
+    this.watchEndCheckMs = watchEndCheckMs;
   }
 
   protected abstract void closeCurrentRequest();
@@ -199,7 +229,8 @@ public abstract class AbstractWatchManager<T extends HasMetadata> implements Wat
     logger.debug("Scheduling reconnect task in {} ms", delay);
 
     synchronized (this) {
-      reconnectAttempt = Utils.schedule(baseOperation.context.getExecutor(), this::reconnect, delay, TimeUnit.MILLISECONDS);
+      reconnectAttempt = Utils.schedule(baseOperation.getOperationContext().getExecutor(), this::reconnect, delay,
+          TimeUnit.MILLISECONDS);
       if (isForceClosed()) {
         cancelReconnect();
       }
@@ -225,7 +256,7 @@ public abstract class AbstractWatchManager<T extends HasMetadata> implements Wat
   }
 
   final long nextReconnectInterval() {
-    return retryIntervalCalculator.nextReconnectInterval();
+    return Math.max(retryAfterSeconds.getAndSet(0) * 1000L, retryIntervalCalculator.nextReconnectInterval());
   }
 
   void resetReconnectAttempts(WatchRequestState state) {
@@ -288,7 +319,7 @@ public abstract class AbstractWatchManager<T extends HasMetadata> implements Wat
 
   @Override
   public void close() {
-    logger.debug("Force closing the watch {}", this);
+    logger.debug("Force closing the watch");
     closeEvent();
     closeRequest();
     cancelReconnect();
@@ -317,6 +348,7 @@ public abstract class AbstractWatchManager<T extends HasMetadata> implements Wat
   }
 
   protected void onMessage(String message, WatchRequestState state) {
+    endErrors.clear();
     if (state.closed.get() || forceClosed.get()) {
       return;
     }
@@ -354,18 +386,62 @@ public abstract class AbstractWatchManager<T extends HasMetadata> implements Wat
   }
 
   protected boolean onStatus(Status status, WatchRequestState state) {
+    endErrors.clear();
     if (state.closed.get()) {
       return true;
     }
     // The resource version no longer exists - this has to be handled by the caller.
-    if (status.getCode() == HTTP_GONE) {
+    if (Integer.valueOf(HTTP_GONE).equals(status.getCode())) {
       close(new WatcherException(status.getMessage(), new KubernetesClientException(status)));
       return true;
     }
 
     logger.error("Error received: {}, will retry", status);
+    // save the after seconds for the retry attempt
+    retryAfterSeconds.set(Optional.ofNullable(status.getDetails()).map(StatusDetails::getRetryAfterSeconds).orElse(0));
     closeRequest();
     return false;
+  }
+
+  void watchEnded(Throwable t, WatchRequestState state) {
+    state.ended.complete(null);
+    if (state != latestRequestState) {
+      // should not happen, but there is already some mitigation logic in the jetty client,
+      // so we'll guard against an erroneous error after the logical close which would cause
+      // a reconnect
+      if (t != null) {
+        logger.debug("Watch error received after the next watch started", t);
+      }
+      return;
+    }
+    if (t instanceof ProtocolException) {
+      // informers could generally try relisting in this case, but for now would need to override the execeptionHandler
+      close(new WatcherException("Could not process Watch response", t));
+      return;
+    }
+    try {
+      if (t != null) {
+        logEndError(t);
+      }
+    } finally {
+      scheduleReconnect(state);
+    }
+  }
+
+  private void logEndError(Throwable t) {
+    int occurrences = endErrors.compute(t.getClass(), (k, v) -> v == null ? 1 : v + 1);
+    if (t instanceof IOException || t instanceof KubernetesClientException) {
+      // could introspect the KubernetesClientException - it may represent something that should have elevated logging
+      if (occurrences > INFO_LOG_CONNECTION_ERRORS) {
+        logger.info("Watch connection error recieved {} times without progress, will reconnect if possible", occurrences, t);
+      } else {
+        logger.debug("Watch connection error, will reconnect if possible", t);
+      }
+    } else if (occurrences > 1) {
+      logger.info("Unknown Watch error received {} times without progress, will reconnect if possible", occurrences, t);
+    } else {
+      logger.debug("Unknown Watch error received, will reconnect if possible", t);
+    }
   }
 
 }