+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/WatcherWebSocketListener.java
@@ -16,7 +16,6 @@
 package io.fabric8.kubernetes.client.dsl.internal;
 
 import io.fabric8.kubernetes.api.model.HasMetadata;
-import io.fabric8.kubernetes.client.WatcherException;
 import io.fabric8.kubernetes.client.dsl.internal.AbstractWatchManager.WatchRequestState;
 import io.fabric8.kubernetes.client.http.WebSocket;
 import org.slf4j.Logger;
@@ -43,13 +42,8 @@ class WatcherWebSocketListener<T extends HasMetadata> implements WebSocket.Liste
   }
 
   @Override
-  public void onError(WebSocket webSocket, Throwable t, boolean connectionError) {
-    if (connectionError) {
-      logger.debug("WebSocket error received", t);
-      manager.scheduleReconnect(state);
-    } else {
-      manager.close(new WatcherException("Could not process websocket message", t));
-    }
+  public void onError(WebSocket webSocket, Throwable t) {
+    manager.watchEnded(t, state);
   }
 
   @Override
@@ -69,8 +63,11 @@ class WatcherWebSocketListener<T extends HasMetadata> implements WebSocket.Liste
   @Override
   public void onClose(WebSocket webSocket, int code, String reason) {
     logger.debug("WebSocket close received. code: {}, reason: {}", code, reason);
-    webSocket.sendClose(code, reason);
-    manager.scheduleReconnect(state);
+    try {
+      webSocket.sendClose(code, reason);
+    } finally {
+      manager.watchEnded(null, state);
+    }
   }
 
 }